fix: error for the status querying eth transactions
